Allow to link the socket in the standard path
The pipewire-0 socket must be at /run/user/XXXX/ path (this is,
at the default $XDG_RUNTIME_DIR folder), but the daemon options
of snapd doesn't allow to do "listen-stream" outside the snap's
$XDG_RUNTIME_DIR folder, which, in the case of the snapped
pipewire is /run/user/XXXX/snap.pipewire/, so a link is
mandatory.

This patch allows to create such a link.