UDENG-434: Add support for audio-playback in snapctl2
Snapctl2 call allows to check from inside a snap if another snap
has connected any of the "cups-control", "audio-record" or
"pulseaudio" interfaces.

For the pipewire support, it would be a good idea to also have
access to the "audio-playback" interface, because, although it
is allowed by default and connected automatically, if an user
manually disconnects it, that wouldn't be honored in pipewire.

This patch does this.