UCD-115: limit access to local wayland sockets
To allow any program to get access to the Wayland socket, no
matter if it is running inside or outside a snap, a hard link is
made from the real socket (located at /run/user/XXXX/snap.ubuntu-desktop-session/wayland-0)
to the expected place (/run/user/XXXX/wayland-0).

Unfortunately, due to the way AppArmor works, it is able to
detect that, although a snapped program is accessing the global
path, it really is trying to access the socket in the local
path, something that, by default, is forbidden.

Currently this is solved with this AppArmor rule:

    owner /run/user/[0-9]*/{,*/}wayland-[0-9]* rw,

This rule has the problem that it allows any snap to access any
wayland-X file/socket located inside the private XDG folder of
any other snap, which is a security breach (for example, it
allows two apparently unrelated snaps to communicate between
them without any kind of control).

The solution presented in this patch consists on allowing to
access only to the socket placed in the folder of the snap that
is implementing the wayland slot.